requirements.txt
fastapi==0.111.0
pandas==2.2.2
Pillow==10.3.0
pydantic==2.7.1
python-dotenv==1.0.1
Requests==2.32.2
uvicorn==0.29.0
